requirements.txt
# Face Detection & Recognition System — Dependencies
# Core ML
facenet-pytorch>=2.5.3
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.16.0

# Computer Vision
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Vector Search
faiss-cpu>=1.7.4

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# Caching (optional — for production)
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
